gestaoMeacnicaBeck-main/requirements.txt
asgiref==3.11.1
brotli==1.2.0
cffi==2.0.0
cssselect2==0.9.0
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
fonttools==4.61.1
inflection==0.5.1
packaging==26.0
pillow==12.1.1
pycparser==3.0
pydyf==0.12.1
PyJWT==2.11.0
pyphen==0.17.2
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
uritemplate==4.2.0
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
psycopg2-binary==2.9.9